reorganiza componentes
